Add really cool locking mechanism support :) Now whenever you open the contact editor, the resource is locked immediately, and as soon as you close the editor, the contact is written back to the backend. So you can add contacts now from KMail without loosing your previous changes...
(Will make the saving timer dependent later...)

svn path=/trunk/kdepim/; revision=246317

+++ kaddressbook/jumpbuttonbar.cpp
@@ -162,7 +162,7 @@ void JumpButtonBar::updateButtons()
         button->show();
         current++;      
       } else {
-        int pos = ( current + offset >= characters.count() ?
+        int pos = ( current + offset >= (int)characters.count() ?
                     characters.count() - 1 : current + offset - 1 );
         JumpButton *button = new JumpButton( characters[ current ],
 			                     characters[ pos ], this );
